+++ src/app/TakeScanner.jsx
@@ -2,6 +2,10 @@ import { useRef } from 'react';
 
 import { openDirectory } from '@/qrcode/DirectoryPicker';
 import { scanVideoBlobForQRCodes } from '@/qrcode/QRCodeReader';
+import { formatExportName } from '@/serdes/ExportNameFormat';
+import { useDocumentStore } from '@/stores/document/use';
+import { useCurrentDocumentId } from '@/stores/user';
+import { downloadText } from '@/utils/Downloader';
 import { extname } from '@/utils/PathHelper';
 
 import { setVideoSrcBlob } from './VideoBlobSource';
@@ -21,6 +25,8 @@ import { setVideoSrcBlob } from './VideoBlobSource';
  * @param {OnScannerChangeCallback} props.onChange
  */
 export default function TakeScanner({ className, onChange }) {
+  const UNSAFE_getStore = useDocumentStore((ctx) => ctx.UNSAFE_getStore);
+  const documentId = useCurrentDocumentId();
   const videoRef = useRef(/** @type {HTMLVideoElement|null} */ (null));
   const inputRef = useRef(/** @type {Record<string, ScannerResult>} */ ({}));
   const outputRef = useRef(/** @type {Record<string, string>} */ ({}));
@@ -84,6 +90,25 @@ export default function TakeScanner({ className, onChange }) {
 
     console.log(`[TakeScanner] Renamed videos.`);
   }
+  async function onExportClick() {
+    let input = inputRef.current;
+
+    let lines = [];
+    for (let key of Object.keys(input)) {
+      lines.push(`${key},${input[key].code}`);
+    }
+    const result = lines.join('\n');
+
+    const store = UNSAFE_getStore();
+    const fileName = formatExportName(
+      store,
+      documentId,
+      '',
+      'SCANNED_NAMES',
+      'csv',
+    );
+    downloadText(fileName, result);
+  }
   return (
     <>
       <button className={className} onClick={onScanClick}>
@@ -94,6 +119,9 @@ export default function TakeScanner({ className, onChange }) {
       </button>
       <button className={className} onClick={onApplyClick}>
         Apply directory
+      </button>
+      <button className={className} onClick={onExportClick}>
+        Export list to .csv
       </button>
       <video
         ref={videoRef}