+++ src/components/settings/SettingsProjectExportZIPButton.jsx
@@ -4,13 +4,13 @@ import DownloadIcon from '@material-symbols/svg-400/rounded/download.svg';
 import { Zip, ZipPassThrough } from 'fflate';
 
 import { getVideoBlob } from '@/recorder/cache';
+import { formatExportName } from '@/serdes/ExportNameFormat';
 import { getDocumentById } from '@/stores/document';
 import { useDocumentStore } from '@/stores/document/use';
 import { getIDBKeyFromTakeId } from '@/stores/document/value';
 import { useCurrentDocumentId } from '@/stores/user';
 import { downloadURLImpl } from '@/utils/Downloader';
 
-import { formatProjectId } from '../takes/TakeNameFormat';
 import SettingsFieldButton from './SettingsFieldButton';
 
 export default function SettingsProjectExportZIPButton() {
@@ -24,18 +24,13 @@ export default function SettingsProjectExportZIPButton() {
       const document = getDocumentById(store, documentId);
       const takes = Object.values(document?.takes || {});
 
-      const projectId = formatProjectId(
-        document?.settings?.projectId || document?.documentTitle,
+      const zippedFileName = formatExportName(
+        store,
+        documentId,
+        '',
+        'EXPORTED_DATA',
+        'zip',
       );
-      const date = new Date();
-      const year = String(date.getFullYear());
-      const month = String(date.getMonth() + 1).padStart(2, '0');
-      const day = String(date.getDate() + 1).padStart(2, '0');
-      const hour = String(date.getHours()).padStart(2, '0');
-      const minute = String(date.getMinutes()).padStart(2, '0');
-      const second = String(date.getSeconds()).padStart(2, '0');
-      const dateString = `${year}_${month}_${day}_${hour}_${minute}_${second}`;
-      const zippedFileName = `${projectId}_EXPORTED_DATA_${dateString}.zip`;
 
       (async () => {
         setDisabled(true);